merge: 合并master任务记录ID溢出修复(26074688c)
master改动:
- queststatus 自增ID由int改long,避免溢出
- 新增幂等BIGINT迁移 V1.11.7__queststatusid_bigint.sql

冲突解决:
- Character.java: master改用rs.getLong(1)内联,HEAD保留questStatusId变量,
  统一为long类型变量+setLong,兼顾可读性与BIGINT修复

+++ gms-server/src/main/java/org/gms/client/Character.java
@@ -8124,11 +8124,11 @@ public class Character extends AbstractCharacterObject {
                             if (!rs.next()) {
                                 throw new SQLException("批量插入任务状态后未返回自增ID");
                             }
-                            int questStatusId = rs.getInt(1);
+                            long questStatusId = rs.getLong(1);
 
                             for (int mob : qs.getProgress().keySet()) {
                                 psProgress.setInt(1, id);
-                                psProgress.setInt(2, questStatusId);
+                                psProgress.setLong(2, questStatusId);
                                 psProgress.setInt(3, mob);
                                 psProgress.setString(4, qs.getProgress(mob));
                                 psProgress.addBatch();
@@ -8136,7 +8136,7 @@ public class Character extends AbstractCharacterObject {
 
                             for (int i = 0; i < qs.getMedalMaps().size(); i++) {
                                 psMedal.setInt(1, id);
-                                psMedal.setInt(2, questStatusId);
+                                psMedal.setLong(2, questStatusId);
                                 psMedal.setInt(3, qs.getMedalMaps().get(i));
                                 psMedal.addBatch();
                             }
